split  out worker and load test functions
